Rename highlighting files to reduce diff
