+++ packages/provider-elasticsearch/src/example-types/results/highlight.js
